refactor: Clean up unused types and misplaced types
